// src/types/api/order.ts
export interface GuestOrderResponse {
  success: boolean;
  message?: string;
  data: {
    order: {
      id: number;
      orderNumber: string;
      status: string;
      paymentStatus: string;
      trackingCode: string;
      consignmentId: number;
      total: number;
      subTotal: number;
      discount: number;
      createdAt: string;
      updatedAt: string;
    };
    customer: {
      name: string;
      email: string;
      phone: string;
    };
    shipping: {
      name: string;
      address: string;
    };
    items: Array<{
      name: string;
      sku: string;
      quantity: number;
      price: number;
      salePrice: number;
      photo: string | null;
    }>;
    paymentMethod: string;
  };
}